+++ src/aui-layout/js/aui-layout-builder.js
@@ -4,14 +4,7 @@
  * @module aui-layout-builder
  */
 
-var RESPONSIVENESS_BREAKPOINT = 992,
-    TPL_LAYOUT_CONTAINER = '<div class="layout-builder-layout-container"></div>';
-
-/**
- * Fired when the the responsive mode changes.
- *
- * @event columnModeChange
- */
+var TPL_LAYOUT_CONTAINER = '<div class="layout-builder-layout-container"></div>';
 
 /**
  * A base class for Layout Builder.
@@ -36,15 +29,6 @@ A.LayoutBuilder = A.Base.create('layout-builder', A.Base, [
 ], {
 
     /**
-     * Determines if columns should collapse.
-     *
-     * @property _isColumnModeEnabled
-     * @type {Boolean}
-     * @protected
-     */
-    _isColumnModeEnabled: null,
-
-    /**
      * The node where the layout will be rendered.
      *
      * @property _layoutContainer
@@ -66,16 +50,14 @@ A.LayoutBuilder = A.Base.create('layout-builder', A.Base, [
         this._createLayoutContainer(container);
 
         layout.addTarget(this);
+
+        this._eventHandles = [
+            this.after('layoutChange', A.bind(this._afterLayoutChange, this))
+        ];
+
         layout.draw(this._layoutContainer);
 
         this._layoutContainer.unselectable();
-
-        this._eventHandles = [
-            this.after('layoutChange', A.bind(this._afterLayoutChange, this)),
-            A.on('windowresize', A.bind(this._afterWindowResize, this))
-        ];
-
-        this._handleResponsive(A.config.win.innerWidth);
     },
 
     /**
@@ -108,19 +90,6 @@ A.LayoutBuilder = A.Base.create('layout-builder', A.Base, [
     },
 
     /**
-     * Fired after window resize.
-     *
-     * @method _afterWindowResize
-     * @param {EventFacade} event
-     * @protected
-     */
-    _afterWindowResize: function(event) {
-        var viewportSize = event.target.get('innerWidth');
-
-        this._handleResponsive(viewportSize);
-    },
-
-    /**
      * Create layout container node.
      *
      * @method _createLayoutContainer
@@ -130,22 +99,6 @@ A.LayoutBuilder = A.Base.create('layout-builder', A.Base, [
     _createLayoutContainer: function(container) {
         this._layoutContainer = A.Node.create(TPL_LAYOUT_CONTAINER);
         container.append(this._layoutContainer);
-    },
-
-    /**
-     * Fires the `columnModeChange` event.
-     *
-     * @method _handleResponsive
-     * @param {Number} viewportSize
-     * @protected
-     */
-    _handleResponsive: function(viewportSize) {
-        var enableColumnMode = viewportSize >= RESPONSIVENESS_BREAKPOINT;
-
-        if (this._isColumnModeEnabled !== enableColumnMode) {
-            this._isColumnModeEnabled = enableColumnMode;
-            this.fire('columnModeChange');
-        }
     }
 }, {
     /**