+++ src/aui-layout/HISTORY.md
@@ -4,6 +4,7 @@
 
 ## @VERSION@
 
+* [AUI-1754](https://issues.liferay.com/browse/AUI-1754) Layout should have option to keep all columns in a row with the same height
 * [AUI-1780](https://issues.liferay.com/browse/AUI-1780) Fix bug with button for adding rows adding multiple rows on each click
 * [AUI-1757](https://issues.liferay.com/browse/AUI-1757) Improvements to layout builder's button for adding rows
 * [AUI-1721](https://issues.liferay.com/browse/AUI-1721) Improve interactions for LayoutBuilder on smartphones